+++ src/main/java/com/faturamento/faturamento_core/domain/dto/produto/ProdutoUpdateDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record ProdutoUpdateDTO (
         @NotBlank(message = "O nome não pode estar em branco")
         String nome,
@@ -13,6 +15,6 @@ public record ProdutoUpdateDTO (
 
         @NotNull(message = "O preço é obrigatório")
         @Positive(message = "O preço deve ser maior que zero")
-        Double preco
+        BigDecimal preco
 ) {
 }
